fix checkpoint dump review blockers

+++ pkg/fileservice/aws_sdk_v2.go
@@ -498,7 +498,7 @@ func (a *AwsSDKv2) WriteMultipartParallel(
 	type partBuffer struct {
 		buf    []byte
 		n      int
-		tokens int
+		tokens int64
 	}
 
 	releasePartBuffer := func(part *partBuffer) {
